fix: moved user subscription into separate domain context
