+++ src/Domain/Handler/SubscriptionReassignedEventHandler.ts
@@ -10,8 +10,8 @@ import TYPES from '../../Bootstrap/Types'
 import { RoleServiceInterface } from '../Role/RoleServiceInterface'
 import { User } from '../User/User'
 import { UserRepositoryInterface } from '../User/UserRepositoryInterface'
-import { UserSubscription } from '../User/UserSubscription'
-import { UserSubscriptionRepositoryInterface } from '../User/UserSubscriptionRepositoryInterface'
+import { UserSubscription } from '../Subscription/UserSubscription'
+import { UserSubscriptionRepositoryInterface } from '../Subscription/UserSubscriptionRepositoryInterface'
 
 @injectable()
 export class SubscriptionReassignedEventHandler
